Move sub-classes of TableLayout class to separate file
