+++ src/System.Windows.Forms/src/System/Windows/Forms/Layout/TableLayout.cs
@@ -5,13 +5,12 @@
 #nullable disable
 
 using System.Collections;
-using System.Collections.Specialized;
 using System.Diagnostics;
 using System.Drawing;
 
 namespace System.Windows.Forms.Layout
 {
-    internal class TableLayout : LayoutEngine
+    internal partial class TableLayout : LayoutEngine
     {
         // This code was taken from ndp\clr\src\BCL\System\Array.cs
         // This changeset replaced the sorting algorithm when elements with the same
@@ -24,128 +23,6 @@ namespace System.Windows.Forms.Layout
         {
             // Note both may be negative, if we are dealing with arrays w/ negative lower bounds.
             return low + ((hi - low) >> 1);
-        }
-
-        // Private value type used by the Sort methods.
-        private struct SorterObjectArray
-        {
-            private readonly object[] keys;
-            private readonly IComparer comparer;
-
-            internal SorterObjectArray(object[] keys, IComparer comparer)
-            {
-                if (comparer is null)
-                {
-                    comparer = Comparer.Default;
-                }
-
-                this.keys = keys;
-                this.comparer = comparer;
-            }
-
-            internal void SwapIfGreaterWithItems(int a, int b)
-            {
-                if (a != b)
-                {
-                    try
-                    {
-                        if (comparer.Compare(keys[a], keys[b]) > 0)
-                        {
-                            object temp = keys[a];
-                            keys[a] = keys[b];
-                            keys[b] = temp;
-                        }
-                    }
-                    catch (IndexOutOfRangeException)
-                    {
-                        throw new ArgumentException();
-                    }
-                    catch (Exception)
-                    {
-                        throw new InvalidOperationException();
-                    }
-                }
-            }
-
-            internal void QuickSort(int left, int right)
-            {
-                // Can use the much faster jit helpers for array access.
-                do
-                {
-                    int i = left;
-                    int j = right;
-
-                    // pre-sort the low, middle (pivot), and high values in place.
-                    // this improves performance in the face of already sorted data, or
-                    // data that is made up of multiple sorted runs appended together.
-                    int middle = GetMedian(i, j);
-                    SwapIfGreaterWithItems(i, middle); // swap the low with the mid point
-                    SwapIfGreaterWithItems(i, j);      // swap the low with the high
-                    SwapIfGreaterWithItems(middle, j); // swap the middle with the high
-
-                    object x = keys[middle];
-                    do
-                    {
-                        // Add a try block here to detect IComparers (or their
-                        // underlying IComparables, etc) that are bogus.
-                        try
-                        {
-                            while (comparer.Compare(keys[i], x) < 0)
-                            {
-                                i++;
-                            }
-
-                            while (comparer.Compare(x, keys[j]) < 0)
-                            {
-                                j--;
-                            }
-                        }
-                        catch (IndexOutOfRangeException)
-                        {
-                            throw new ArgumentException();
-                        }
-                        catch (Exception)
-                        {
-                            throw new InvalidOperationException();
-                        }
-
-                        if (i > j)
-                        {
-                            break;
-                        }
-
-                        if (i < j)
-                        {
-                            object key = keys[i];
-                            keys[i] = keys[j];
-                            keys[j] = key;
-                        }
-
-                        i++;
-                        j--;
-                    }
-                    while (i <= j);
-                    if (j - left <= right - i)
-                    {
-                        if (left < j)
-                        {
-                            QuickSort(left, j);
-                        }
-
-                        left = i;
-                    }
-                    else
-                    {
-                        if (i < right)
-                        {
-                            QuickSort(i, right);
-                        }
-
-                        right = j;
-                    }
-                }
-                while (left < right);
-            }
         }
 
         private static void Sort(object[] array, IComparer comparer)
@@ -1514,84 +1391,6 @@ namespace System.Windows.Forms.Layout
             Debug.Assert(GetLayoutInfo(element) == value, "GetLayoutInfo should return the same value as we set it to");
         }
 
-        /// <summary>
-        ///  This class contains layout related information pertaining to a child control of the
-        ///  container being laid out. It contains Row,column assignments as well as RowSpan/ColumnSpan.
-        ///  This class is used from ContainerInfo as a way of caching information about child controls.
-        /// </summary>
-        internal sealed class LayoutInfo
-        {
-            public LayoutInfo(IArrangedElement element)
-            {
-                Element = element;
-            }
-
-            public bool IsAbsolutelyPositioned => RowPosition >= 0 && ColumnPosition >= 0;
-
-            public IArrangedElement Element { get; }
-
-            /// <summary>
-            ///  Corresponds to TableLayoutSettings.SetRow. Can be -1 indicating that it is a
-            ///  "flow" element and will fit in as necessary. This occurs when a control is
-            ///  just added without specific position.
-            /// </summary>
-            public int RowPosition { get; set; } = -1;
-
-            /// <summary>
-            ///  Corresponds to TableLayoutSettings.SetColumn. Can be -1 indicating that it is a
-            ///  "flow" element and will fit in as necessary. This occurs when a control is
-            ///  just added without specific position.
-            /// </summary>
-            public int ColumnPosition { get; set; } = -1;
-
-            public int RowStart { get; set; } = -1;
-
-            public int ColumnStart { get; set; } = -1;
-
-            public int ColumnSpan { get; set; } = 1;
-
-            public int RowSpan { get; set; } = 1;
-
-#if DEBUG
-            public LayoutInfo Clone() => new LayoutInfo(Element)
-            {
-                RowStart = RowStart,
-                ColumnStart = ColumnStart,
-                RowSpan = RowSpan,
-                ColumnSpan = ColumnSpan,
-                RowPosition = RowPosition,
-                ColumnPosition = ColumnPosition
-            };
-
-            public override bool Equals(object obj)
-            {
-                if (!(obj is LayoutInfo other))
-                {
-                    return false;
-                }
-
-                return other.RowStart == RowStart
-                    && other.ColumnStart == ColumnStart
-                    && other.RowSpan == RowSpan
-                    && other.ColumnSpan == ColumnSpan
-                    && other.RowPosition == RowPosition
-                    && other.ColumnPosition == ColumnPosition;
-            }
-
-            public override int GetHashCode()
-            {
-                var hash = new HashCode();
-                hash.Add(RowStart);
-                hash.Add(ColumnStart);
-                hash.Add(RowSpan);
-                hash.Add(ColumnSpan);
-                hash.Add(RowPosition);
-                hash.Add(ColumnPosition);
-                return hash.ToHashCode();
-            }
-#endif
-        }
-
         #region ContainerInfo
         internal static bool HasCachedAssignments(ContainerInfo containerInfo)
         {
@@ -1617,657 +1416,7 @@ namespace System.Windows.Forms.Layout
 
             return containerInfo;
         }
-
-        /// <summary>
-        ///  this class contains layout related information pertaining to the container
-        ///  being laid out by this instance of the TableLayout.  It contains references
-        ///  to all the information that should be used from the table layout engine,
-        ///  as this class is responsible for caching information about the control and
-        ///  it's children being layed out.
-        /// </summary>
-        internal sealed class ContainerInfo
-        {
-            private static readonly Strip[] emptyStrip = Array.Empty<Strip>();
-
-            private static readonly int stateValid = BitVector32.CreateMask();
-            private static readonly int stateChildInfoValid = BitVector32.CreateMask(stateValid);
-            private static readonly int stateChildHasColumnSpan = BitVector32.CreateMask(stateChildInfoValid);
-            private static readonly int stateChildHasRowSpan = BitVector32.CreateMask(stateChildHasColumnSpan);
-
-            private int _cellBorderWidth;  //the width for the cell border
-            private Strip[] _cols = emptyStrip;
-            private Strip[] _rows = emptyStrip;
-            private int _maxRows;
-            private int _maxColumns;
-            private TableLayoutRowStyleCollection _rowStyles;
-            private TableLayoutColumnStyleCollection _colStyles;
-            private TableLayoutPanelGrowStyle _growStyle;
-            private readonly IArrangedElement _container;
-            private LayoutInfo[] _childInfo;
-            private int _countFixedChildren;
-            private int _minRowsAndColumns; // The minimum space required to put all the controls without overlapping
-            private int _minColumns; // The minimum number of columns required in order to put all absolutely positioned control on the table
-            private int _minRows; // The minimum number of rows required in order to put all absolutely positioned control on the table
-
-            private BitVector32 _state;
-
-            public ContainerInfo(IArrangedElement container)
-            {
-                _container = container;
-                _growStyle = TableLayoutPanelGrowStyle.AddRows;
-            }
-
-            public ContainerInfo(ContainerInfo containerInfo)
-            {
-                _cellBorderWidth = containerInfo.CellBorderWidth;
-                _maxRows = containerInfo.MaxRows;
-                _maxColumns = containerInfo.MaxColumns;
-                _growStyle = containerInfo.GrowStyle;
-                _container = containerInfo.Container;
-                _rowStyles = containerInfo.RowStyles;
-                _colStyles = containerInfo.ColumnStyles;
-            }
-
-            /// <summary>
-            ///  the container being laid out
-            /// </summary>
-            public IArrangedElement Container
-            {
-                get { return _container; }
-            }
-
-            public int CellBorderWidth
-            {
-                get { return _cellBorderWidth; }
-                set { _cellBorderWidth = value; }
-            }
-
-            /// <summary>
-            ///  list of ints that represent the sizes of individual columns
-            /// </summary>
-            public Strip[] Columns
-            {
-                get { return _cols; }
-                set
-                {
-                    Debug.Assert(_cols.Length != value.Length, "PERF: should not allocate strips, we've already got an array");
-                    _cols = value;
-                }
-            }
-
-            ///
-            ///  list of ints that represent the sizes of individual rows
-            ///
-            public Strip[] Rows
-            {
-                get { return _rows; }
-
-                set
-                {
-                    Debug.Assert(_rows.Length != value.Length, "PERF: should not allocate strips, we've already got an array");
-                    _rows = value;
-                }
-            }
-
-            /// <summary>
-            ///  Same as TableLayoutSettings.RowCount
-            /// </summary>
-            public int MaxRows
-            {
-                get { return _maxRows; }
-                set
-                {
-                    if (_maxRows != value)
-                    {
-                        _maxRows = value;
-
-                        //invalidate the cache whenever we change the number of rows
-                        Valid = false;
-                    }
-                }
-            }
-
-            /// <summary>
-            ///  Same as TableLayoutSettings.ColumnCount
-            /// </summary>
-            public int MaxColumns
-            {
-                get { return _maxColumns; }
-
-                set
-                {
-                    if (_maxColumns != value)
-                    {
-                        _maxColumns = value;
-
-                        //invalidate the cache whenever we change the number of columns
-                        Valid = false;
-                    }
-                }
-            }
-
-            ///  Cached information
-            public int MinRowsAndColumns
-            {
-                get
-                {
-                    Debug.Assert(ChildInfoValid, "Fetching invalid information");
-                    return _minRowsAndColumns;
-                }
-            }
-
-            ///  Cached information
-            public int MinColumns
-            {
-                get
-                {
-                    Debug.Assert(ChildInfoValid, "Fetching invalid information");
-                    return _minColumns;
-                }
-            }
-
-            ///  Cached information
-            public int MinRows
-            {
-                get
-                {
-                    Debug.Assert(ChildInfoValid, "Fetching invalid information");
-                    return _minRows;
-                }
-            }
-
-            /// <summary>
-            ///  Gets/sets the grow style for our containerinfo.  This
-            ///  is used to determine if we will add rows/cols/or throw
-            ///  when the table gets full.
-            /// </summary>
-            public TableLayoutPanelGrowStyle GrowStyle
-            {
-                get
-                {
-                    return _growStyle;
-                }
-                set
-                {
-                    if (_growStyle != value)
-                    {
-                        _growStyle = value;
-                        Valid = false; // throw away cached row and column assignments
-                    }
-                }
-            }
-
-            public TableLayoutRowStyleCollection RowStyles
-            {
-                get
-                {
-                    if (_rowStyles is null)
-                    {
-                        _rowStyles = new TableLayoutRowStyleCollection(_container);
-                    }
-
-                    return _rowStyles;
-                }
-                set
-                {
-                    _rowStyles = value;
-                    if (_rowStyles != null)
-                    {
-                        _rowStyles.EnsureOwnership(_container);
-                    }
-                }
-            }
-
-            public TableLayoutColumnStyleCollection ColumnStyles
-            {
-                get
-                {
-                    if (_colStyles is null)
-                    {
-                        _colStyles = new TableLayoutColumnStyleCollection(_container);
-                    }
-
-                    return _colStyles;
-                }
-                set
-                {
-                    _colStyles = value;
-                    if (_colStyles != null)
-                    {
-                        _colStyles.EnsureOwnership(_container);
-                    }
-                }
-            }
-
-            /// <summary>
-            ///  gets cached information about the children of the control being layed out.
-            /// </summary>
-            public LayoutInfo[] ChildrenInfo
-            {
-                get
-                {
-                    if (!_state[stateChildInfoValid])
-                    {
-                        _countFixedChildren = 0;
-
-                        _minRowsAndColumns = 0;
-                        _minColumns = 0;
-                        _minRows = 0;
-                        ArrangedElementCollection children = Container.Children;
-                        LayoutInfo[] childInfo = new LayoutInfo[children.Count];
-                        int nonParticipatingElements = 0;
-                        int index = 0;
-                        for (int i = 0; i < children.Count; i++)
-                        {
-                            IArrangedElement element = children[i];
-                            if (!element.ParticipatesInLayout)
-                            {
-                                // If the element does not participate in layout (i.e., Visible = false), we
-                                // exclude it from the childrenInfos list and it is ignored by the engine.
-                                nonParticipatingElements++;
-                                continue;
-                            }
-
-                            LayoutInfo layoutInfo = GetLayoutInfo(element);
-                            if (layoutInfo.IsAbsolutelyPositioned)
-                            {
-                                _countFixedChildren++;
-                            }
-
-                            childInfo[index++] = layoutInfo;
-                            _minRowsAndColumns += layoutInfo.RowSpan * layoutInfo.ColumnSpan;
-                            if (layoutInfo.IsAbsolutelyPositioned)
-                            {
-                                _minColumns = Math.Max(_minColumns, layoutInfo.ColumnPosition + layoutInfo.ColumnSpan);
-                                _minRows = Math.Max(_minRows, layoutInfo.RowPosition + layoutInfo.RowSpan);
-                            }
-                        }
-
-                        // shorten the array if necessary.
-                        if (nonParticipatingElements > 0)
-                        {
-                            LayoutInfo[] trimmedChildInfo = new LayoutInfo[childInfo.Length - nonParticipatingElements];
-                            Array.Copy(childInfo, trimmedChildInfo, trimmedChildInfo.Length);
-                            _childInfo = trimmedChildInfo;
-                        }
-                        else
-                        {
-                            _childInfo = childInfo;
-                        }
-
-                        _state[stateChildInfoValid] = true;
-                    }
-
-                    return _childInfo ?? (Array.Empty<LayoutInfo>());
-                }
-            }
-
-            public bool ChildInfoValid
-            {
-                get { return _state[stateChildInfoValid]; }
-            }
-
-            public LayoutInfo[] FixedChildrenInfo
-            {
-                get
-                {
-                    Debug.Assert(ChildInfoValid, "Fetched invalid information");
-                    // we only get this in a cached scenario - so we don't have to worry about caching it.
-                    LayoutInfo[] fixedChildren = new LayoutInfo[_countFixedChildren];
-                    if (HasChildWithAbsolutePositioning)
-                    {
-                        int index = 0;
-                        for (int i = 0; i < _childInfo.Length; i++)
-                        {
-                            if (_childInfo[i].IsAbsolutelyPositioned)
-                            {
-                                fixedChildren[index++] = _childInfo[i];
-                            }
-                        }
-
-                        Sort(fixedChildren, PreAssignedPositionComparer.GetInstance);
-                    }
-
-                    return fixedChildren;
-                }
-            }
-
-            public bool Valid
-            {
-                get { return _state[stateValid]; }
-                set
-                {
-                    _state[stateValid] = value;
-                    if (!_state[stateValid])
-                    {
-                        _state[stateChildInfoValid] = false;
-                    }
-                }
-            }
-
-            public bool HasChildWithAbsolutePositioning
-            {
-                get { return _countFixedChildren > 0; }
-            }
-
-            public bool HasMultiplePercentColumns
-            {
-                get
-                {
-                    if (_colStyles != null)
-                    {
-                        bool foundAny = false;
-                        foreach (ColumnStyle style in _colStyles)
-                        {
-                            if (style.SizeType == SizeType.Percent)
-                            {
-                                if (foundAny)
-                                {
-                                    return true;
-                                }
-
-                                foundAny = true;
-                            }
-                        }
-                    }
-
-                    return false;
-                }
-            }
-
-            public bool ChildHasColumnSpan
-            {
-                get { return _state[stateChildHasColumnSpan]; }
-                set { _state[stateChildHasColumnSpan] = value; }
-            }
-
-            public bool ChildHasRowSpan
-            {
-                get { return _state[stateChildHasRowSpan]; }
-                set { _state[stateChildHasRowSpan] = value; }
-            }
-
-            public Size GetCachedPreferredSize(Size proposedConstraints, out bool isValid)
-            {
-                isValid = false;
-                if (proposedConstraints.Height == 0 || proposedConstraints.Width == 0)
-                {
-                    Size cachedSize = CommonProperties.xGetPreferredSizeCache(Container);
-
-                    if (!cachedSize.IsEmpty)
-                    {
-                        isValid = true;
-                        return cachedSize;
-                    }
-                }
-
-                return Size.Empty;
-            }
-        }
         #endregion
-
-        #region SizeProxy implementations
-        //sizeProxy. Takes a strip and return its minSize or maxSize accordingly
-        private abstract class SizeProxy
-        {
-            protected Strip strip;
-            public Strip Strip
-            {
-                get { return strip; }
-                set { strip = value; }
-            }
-
-            public abstract int Size
-            {
-                get; set;
-            }
-        }
-
-        private class MinSizeProxy : SizeProxy
-        {
-            private static readonly MinSizeProxy instance = new MinSizeProxy();
-            public override int Size
-            {
-                get { return strip.MinSize; }
-                set { strip.MinSize = value; }
-            }
-
-            public static MinSizeProxy GetInstance
-            {
-                get
-                {
-                    return instance;
-                }
-            }
-        }
-
-        private class MaxSizeProxy : SizeProxy
-        {
-            private static readonly MaxSizeProxy instance = new MaxSizeProxy();
-            public override int Size
-            {
-                get { return strip.MaxSize; }
-                set { strip.MaxSize = value; }
-            }
-
-            public static MaxSizeProxy GetInstance
-            {
-                get
-                {
-                    return instance;
-                }
-            }
-        }
-        #endregion
-
-        #region ICompare implementations
-        private abstract class SpanComparer : IComparer
-        {
-            public abstract int GetSpan(LayoutInfo layoutInfo);
-
-            public int Compare(object x, object y)
-            {
-                LayoutInfo xInfo = (LayoutInfo)x;
-                LayoutInfo yInfo = (LayoutInfo)y;
-                return GetSpan(xInfo) - GetSpan(yInfo);
-            }
-        }
-
-        private class RowSpanComparer : SpanComparer
-        {
-            private static readonly RowSpanComparer instance = new RowSpanComparer();
-
-            public override int GetSpan(LayoutInfo layoutInfo)
-            {
-                return layoutInfo.RowSpan;
-            }
-
-            public static RowSpanComparer GetInstance
-            {
-                get { return instance; }
-            }
-        }
-
-        private class ColumnSpanComparer : SpanComparer
-        {
-            private static readonly ColumnSpanComparer instance = new ColumnSpanComparer();
-
-            public override int GetSpan(LayoutInfo layoutInfo)
-            {
-                return layoutInfo.ColumnSpan;
-            }
-
-            public static ColumnSpanComparer GetInstance
-            {
-                get { return instance; }
-            }
-        }
-
-        private class PostAssignedPositionComparer : IComparer
-        {
-            private static readonly PostAssignedPositionComparer instance = new PostAssignedPositionComparer();
-
-            public static PostAssignedPositionComparer GetInstance
-            {
-                get { return instance; }
-            }
-
-            public int Compare(object x, object y)
-            {
-                LayoutInfo xInfo = (LayoutInfo)x;
-                LayoutInfo yInfo = (LayoutInfo)y;
-                if (xInfo.RowStart < yInfo.RowStart)
-                {
-                    return -1;
-                }
-
-                if (xInfo.RowStart > yInfo.RowStart)
-                {
-                    return 1;
-                }
-
-                if (xInfo.ColumnStart < yInfo.ColumnStart)
-                {
-                    return -1;
-                }
-
-                if (xInfo.ColumnStart > yInfo.ColumnStart)
-                {
-                    return 1;
-                }
-
-                return 0;
-            }
-        }
-
-        private class PreAssignedPositionComparer : IComparer
-        {
-            private static readonly PreAssignedPositionComparer instance = new PreAssignedPositionComparer();
-
-            public static PreAssignedPositionComparer GetInstance
-            {
-                get { return instance; }
-            }
-
-            public int Compare(object x, object y)
-            {
-                LayoutInfo xInfo = (LayoutInfo)x;
-                LayoutInfo yInfo = (LayoutInfo)y;
-                if (xInfo.RowPosition < yInfo.RowPosition)
-                {
-                    return -1;
-                }
-
-                if (xInfo.RowPosition > yInfo.RowPosition)
-                {
-                    return 1;
-                }
-
-                if (xInfo.ColumnPosition < yInfo.ColumnPosition)
-                {
-                    return -1;
-                }
-
-                if (xInfo.ColumnPosition > yInfo.ColumnPosition)
-                {
-                    return 1;
-                }
-
-                return 0;
-            }
-        }
-        #endregion
-
-        #region ReservationGrid
-        // The ReservationGrid is used to track elements which span rows to prevent overlap.
-        private sealed class ReservationGrid
-        {
-            int _numColumns = 1;
-            readonly ArrayList _rows = new ArrayList();
-
-            public bool IsReserved(int column, int rowOffset)
-            {
-                if (rowOffset >= _rows.Count)
-                {
-                    return false;
-                }
-
-                if (column >= ((BitArray)_rows[rowOffset]).Length)
-                {
-                    return false;
-                }
-
-                return ((BitArray)_rows[rowOffset])[column];
-            }
-
-            public void Reserve(int column, int rowOffset)
-            {
-                Debug.Assert(!IsReserved(column, rowOffset), "we should not be reserving already reserved space.");
-                while (rowOffset >= _rows.Count)
-                {
-                    _rows.Add(new BitArray(_numColumns));
-                }
-
-                //increase the length of the _rows[rowOffset] if necessary
-                if (column >= ((BitArray)_rows[rowOffset]).Length)
-                {
-                    ((BitArray)_rows[rowOffset]).Length = column + 1;
-                    if (column >= _numColumns)
-                    {
-                        _numColumns = column + 1;
-                    }
-                }
-
-                ((BitArray)_rows[rowOffset])[column] = true;
-                Debug.Assert(IsReserved(column, rowOffset), "IsReserved/Reserved mismatch.");
-            }
-
-            //reserve all spaces taken by layoutInfo.Element, up till colStop
-            public void ReserveAll(LayoutInfo layoutInfo, int rowStop, int colStop)
-            {
-                for (int rowOffset = 1; rowOffset < rowStop - layoutInfo.RowStart; rowOffset++)
-                {
-                    for (int reservedCol = layoutInfo.ColumnStart; reservedCol < colStop; reservedCol++)
-                    {
-                        Reserve(reservedCol, rowOffset);
-                    }
-                }
-            }
-
-            public void AdvanceRow()
-            {
-                if (_rows.Count > 0)
-                {
-                    _rows.RemoveAt(0);
-                }
-            }
-        }
-        #endregion ReservationGrid
-
-        internal struct Strip
-        {
-            private int _maxSize;
-            private int _minSize;
-            private bool _isStart;  //whether there is an element starting in this strip
-
-            public int MinSize
-            {
-                get { return _minSize; }
-                set { _minSize = value; }
-            }
-
-            public int MaxSize
-            {
-                get { return _maxSize; }
-                set { _maxSize = value; }
-            }
-
-            public bool IsStart
-            {
-                get { return _isStart; }
-                set { _isStart = value; }
-            }
-        }
 
         #region DEBUG
         // Verify that the Row/Column assignments on the control are current.
